+++ crates/nostr/src/event/id.rs
@@ -17,7 +17,7 @@ use super::error::Error;
 use super::{Kind, Tag};
 use crate::nips::nip13;
 use crate::nips::nip19::FromBech32;
-use crate::nips::nip21::NostrURI;
+use crate::nips::nip21::FromNostrUri;
 use crate::util::hex;
 use crate::{PublicKey, Timestamp};