+++ crates/nostr/src/key/public_key.rs
@@ -15,7 +15,7 @@ use serde::{Deserialize, Deserializer, Serialize};
 
 use super::Error;
 use crate::nips::nip19::FromBech32;
-use crate::nips::nip21::NostrURI;
+use crate::nips::nip21::FromNostrUri;
 use crate::util::hex;
 
 /// Public Key